tests/test_perception.py
"""
MediPilot 感知层单元测试
"""
import pytest
from PIL import Image
import numpy as np
from medipilot.perception.screen import Perception, PerceptionError

class TestPerception:
    """感知层测试类"""
    
    @pytest.fixture
    def perception(self):
        """创建感知层实例"""
        return Perception()
    
    @pytest.fixture
    def sample_image(self):
        """创建测试图像"""
        return Image.new('RGB', (800, 600), color='white')
    
    def test_initialization(self, perception):
        """测试初始化"""
        assert perception is not None
        assert hasattr(perception, 'sct')
    
    def test_capture(self, perception):
        """测试截屏功能"""
        img = perception.capture()
        assert isinstance(img, Image.Image)
        assert img.size[0] > 0
        assert img.size[1] > 0
        assert img.mode == 'RGB'
    
    def test_privacy_filter(self, perception, sample_image):
        """测试隐私过滤"""
        filtered = perception.privacy_filter(sample_image)
        assert isinstance(filtered, Image.Image)
        assert filtered.size == sample_image.size
        assert filtered.mode == 'RGB'
    
    def test_privacy_filter_modifies_region(self, perception):
        """测试隐私过滤确实修改了指定区域"""
        # 创建带有文字的图像
        img = Image.new('RGB', (800, 600), color='white')
        from PIL import ImageDraw
        draw = ImageDraw.Draw(img)
        draw.text((50, 50), "SENSITIVE DATA", fill='black')
        
        # 应用隐私过滤
        filtered = perception.privacy_filter(img)
        
        # 检查隐私区域是否被修改
        original_array = np.array(img)
        filtered_array = np.array(filtered)
        
        # 隐私区域应该有差异
        from configs.settings import config
        y1, y2, x1, x2 = config.PRIVACY_REGION
        roi_original = original_array[y1:y2, x1:x2]
        roi_filtered = filtered_array[y1:y2, x1:x2]
        
        # 计算差异
        diff = np.abs(roi_original.astype(float) - roi_filtered.astype(float))
        mean_diff = np.mean(diff)
        
        # 模糊后应该有明显差异（对于有内容的区域）
        assert mean_diff >= 0  # 至少不会完全相同
    
    def test_privacy_filter_with_invalid_image(self, perception):
        """测试无效图像的处理"""
        # 传入None应该抛出异常或返回None
        try:
            result = perception.privacy_filter(None)
            # 如果不抛出异常，至少应该返回None或原值
        except (AttributeError, TypeError):
            # 预期的异常
            pass
    
    def test_som_overlay(self, perception, sample_image):
        """测试SoM网格叠加"""
        marked = perception.apply_som_overlay(sample_image, grid_size=80)
        assert isinstance(marked, Image.Image)
        assert marked.size == sample_image.size
    
    def test_som_overlay_with_different_grid_sizes(self, perception, sample_image):
        """测试不同网格大小"""
        for grid_size in [40, 80, 120]:
            marked = perception.apply_som_overlay(sample_image, grid_size=grid_size)
            assert isinstance(marked, Image.Image)
            assert marked.size == sample_image.size
    
    def test_som_overlay_modifies_image(self, perception, sample_image):
        """测试SoM确实修改了图像"""
        marked = perception.apply_som_overlay(sample_image, grid_size=80)
        
        # 转换为数组比较
        original_array = np.array(sample_image)
        marked_array = np.array(marked)
        
        # 应该有差异（因为添加了网格和文字）
        assert not np.array_equal(original_array, marked_array)
    
    def test_som_overlay_with_invalid_grid_size(self, perception, sample_image):
        """测试无效的网格大小"""
        # 负数应该使用默认值
        marked = perception.apply_som_overlay(sample_image, grid_size=-10)
        assert isinstance(marked, Image.Image)
        
        # 过大的值应该使用默认值
        marked = perception.apply_som_overlay(sample_image, grid_size=10000)
        assert isinstance(marked, Image.Image)
    
    def test_complete_pipeline(self, perception):
        """测试完整的处理流程"""
        # 捕获 -> 脱敏 -> 叠加网格
        img = perception.capture()
        filtered = perception.privacy_filter(img)
        marked = perception.apply_som_overlay(filtered)
        
        assert isinstance(marked, Image.Image)
        assert marked.mode == 'RGB'


if __name__ == "__main__":
    pytest.main([__file__, "-v"])