requirements.txt
# Core Dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# GUI & Automation
pyautogui>=0.9.54
mss>=9.0.1

# Image Processing
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Testing (Development)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Type Checking (Development)
mypy>=1.5.0
types-Pillow>=10.0.0

# Code Quality (Development)
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0

# Optional: For local PII detection models
# ultralytics>=8.0.0  # If using YOLO
# paddleocr>=2.7.0    # If using PaddleOCR
